refactor(core): jsonl.rs → jsonl/ module — walk / liveness / unclaim / health
Behavior-preserving geometry split of the 3274-line source/jsonl.rs into a
directory module; code moved verbatim (95 fns, 39 tests, 401 doc lines —
parity-checked against the original):

- mod.rs      JsonlWatcher + builders/test seams + the run() select loop +
              WatchCtx/SourceDecoders + the per-source fn-pointer type aliases;
              public paths preserved via re-exports (ProbeSnapshot/
              LivenessProbe from liveness, ChildEndUnclaims from unclaim,
              pub(crate) is_subagent_path from walk)
- walk.rs     walk_jsonl + the unified first-sight gate (should_seed_at_eof,
              check_session_ended) + scan_root + emit_first_sight +
              read_head_cwd/read_tail + the oversized handling
              (scan_pending_tasks, TASK_SCAN_BYTES) + detect_parent_id/
              extract_cwd/is_subagent_path
- liveness.rs the whole probe ladder in one home: ProbeSnapshot/LivenessProbe,
              refresh_probe_snapshot (pid_bindings/rebind migration),
              NegativeVouch, emit_session_exit/unbind_session,
              revouch_gated_files, emit_proof_of_life, probe_admits
- unclaim.rs  ChildEndUnclaims + drain_child_end_unclaims (#246)
- health.rs   FailureLatch
- tests.rs    the sibling unit-test module (repo sibling-file convention),
              all 39 tests verbatim — shared helpers (walk_once*, fixtures)
              kept together rather than duplicated per submodule

Only `use` paths, `mod` declarations, and minimal pub(super) visibility bumps
changed (cross-submodule callers + the tests module). Comment-only path
references updated (CLAUDE.md layout/sharp edges, reducer.rs, watcher.rs).
